Refactor JUnit stream parser extraction
